% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.character}
\alias{tson.character}
\title{Make a tson character}
\usage{
tson.character(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson character
}
\description{
Make a tson character
}

