% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.float32.vec}
\alias{tson.float32.vec}
\title{Make a tson float32 vector}
\usage{
tson.float32.vec(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson float32 vector
}
\description{
Make a tson float32 vector
}

