% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.int16.vec}
\alias{tson.int16.vec}
\title{Make a tson int16 vector}
\usage{
tson.int16.vec(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson int16 vector
}
\description{
Make a tson int16 vector
}

