% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.int64.vec}
\alias{tson.int64.vec}
\title{Make a tson int16 vector}
\usage{
tson.int64.vec(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson int64 vector
}
\description{
Make a tson int16 vector
}
