% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.int8.vec}
\alias{tson.int8.vec}
\title{Make a tson int8 vector}
\usage{
tson.int8.vec(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson int8 vector
}
\description{
Make a tson int8 vector
}
