% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.scalar}
\alias{tson.scalar}
\title{Make a tson scalar (ie: singleton)}
\usage{
tson.scalar(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson scalar
}
\description{
Make a tson scalar (ie: singleton)
}

