% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.uint16.vec}
\alias{tson.uint16.vec}
\title{Make a tson uint16 vector}
\usage{
tson.uint16.vec(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson uint16 vector
}
\description{
Make a tson uint16 vector
}
