% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tson.uint32.vec}
\alias{tson.uint32.vec}
\title{Make a tson uint32 vector}
\usage{
tson.uint32.vec(object)
}
\arguments{
\item{object}{A vector or list}
}
\value{
A tson uint32 vector
}
\description{
Make a tson uint32 vector
}

